identity.export: Fix exporting of contact's transports
In the past, the core always put its own list of relay servers as
transports, regardless if we were exporting a local identity, or just a
contact. Now it always encodes the transports as specified in the struct
'id', which gives us proper behaviour regarless if we export local
identity or a remote contact.