Continued with stateless friend requests.
This implementation retains the old behaviour of the friend request
connection, but adds the feature that the friend request is added to
contact database even if remote party does not accept the friendreq
during the lifetime of the friend request connection.
The party sending the friend request flaggs the new friend as
"unconfirmed", and sets "do not connect" flag.

If the friend request connection is closed and we have not gotten any
(positive) answer, we remove the "do not connect" flag, and
then we start "hammering" the potential friend in
hope that it might some day accept the friend request.

If the party, which we friend requested, accepts after the friend
request connection has already died, then we will issue
friendRequesteeAccepted signal, and remove unconfirmedFriendRequest
flag.